% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ssp.R
\name{cosine.sim}
\alias{cosine.sim}
\title{Calculate cosine similarity between two matrices}
\usage{
cosine.sim(x, y)
}
\arguments{
\item{x}{Numeric matrix (genes x samples).}

\item{y}{Numeric matrix (genes x references).}
}
\value{
Cosine similarity matrix (samples x references).
}
\description{
Calculate cosine similarity between two matrices
}
