% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ssp.R
\name{dq.vect}
\alias{dq.vect}
\title{calculate iCMS on bulk data (microarray / RNA-seq) using single-sample predictor}
\usage{
dq.vect(v, min.cor, min.dist, q, metric, allgenes, min.genes)
}
\arguments{
\item{v}{input vector to calculate.}

\item{min.cor}{the minimum correlation threshold to determine the confidence}

\item{min.dist}{the minimum distance to determine the iCMS call}

\item{q}{The quantile that is used to judge distance. q=0.9-1 appears better for non batch and corrected data while q=0.6-0.7 appears better for batch-corrected}

\item{allgenes}{whether to use the extended genes or not, basically only epithelial genes are used}

\item{min.genes}{Minimum number of genes that can calculate iCMS call}

\item{matric}{The method to calculate correlation matrix. Kendall correlation (non parametric) seems more robust but is slow. Pearson is faster, and often good enough.}
}
\value{
iCMS classification for this sample
}
\description{
calculate iCMS on bulk data (microarray / RNA-seq) using single-sample predictor
}
