% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icms.ssc.R
\name{iCMS.KNN}
\alias{iCMS.KNN}
\title{calculate iCMS on bulk data (microarray / RNA-seq) using k-nearest neighbors algorithm}
\usage{
iCMS.KNN(
  ivect,
  nn = 10,
  metric = "kendall",
  allgenes = FALSE,
  min.genes = 30,
  jobs = 4,
  verbose = FALSE
)
}
\arguments{
\item{ivect}{input matrix to calculate. The row will be the gene symbol, and column will be sample name}

\item{nn}{the minimum correlation threshold to determine the confidence (default = 0.1)}

\item{allgenes}{whether to use the extended genes or not, basically only epithelial genes are used (default = FALSE)}

\item{min.genes}{Minimum number of genes that can calculate iCMS call (default = 30 genes)}

\item{jobs}{If parallel execution fails you can try jobs=1 which will use serial apply (default = 4)}

\item{matric}{The method to calculate correlation matrix (default = kendall). Kendall correlation (non parametric) seems more robust but is slow. Pearson is faster, and often good enough.}
}
\value{
iCMS classification
}
\description{
calculate iCMS on bulk data (microarray / RNA-seq) using k-nearest neighbors algorithm
}
