% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icms.ssc.R
\name{iCMS.NTP}
\alias{iCMS.NTP}
\title{calculate iCMS on bulk data (microarray / RNA-seq) using k-nearest neighbors algorithm}
\usage{
iCMS.NTP(ivect, metric = "kendall", bs.iter = 100, jobs = 8)
}
\arguments{
\item{ivect}{input matrix to calculate. The row will be the gene symbol, and column will be sample name}

\item{bs.iter}{the number of iteration to calculate P-value}

\item{jobs}{If parallel execution fails you can try jobs=1 which will use serial apply (default = 4)}

\item{matric}{The method to calculate correlation matrix (default = kendall). Kendall correlation (non parametric) seems more robust but is slow. Pearson is faster, and often good enough.}

\item{allgenes}{whether to use the extended genes or not, basically only epithelial genes are used (default = FALSE)}

\item{min.genes}{Minimum number of genes that can calculate iCMS call (default = 30 genes)}
}
\value{
iCMS classification
}
\description{
calculate iCMS on bulk data (microarray / RNA-seq) using k-nearest neighbors algorithm
}
