% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icms.ssc.R
\name{iCMS.SSC}
\alias{iCMS.SSC}
\title{calculate iCMS on bulk data (microarray / RNA-seq)}
\usage{
iCMS.SSC(
  ivect,
  min.cor = 0.1,
  min.dist = 0.05,
  q = 0.9,
  metric = "kendall",
  allgenes = FALSE,
  min.genes = 30,
  jobs = 8
)
}
\arguments{
\item{ivect}{input matrix to calculate. The row will be the gene symbol, and column will be sample name}

\item{min.cor}{the minimum correlation threshold to determine the confidence (default = 0.1)}

\item{min.dist}{the minimum distance to determine the iCMS call (default = 0.05)}

\item{q}{The quantile that is used to judge distance}

\item{allgenes}{description}

\item{min.genes}{Minimum number of genes that can calculate iCMS call (default = 30 genes)}

\item{matric}{The method to calculate correlation matrix (default = kendall)}
}
\value{
iCMS classification
}
\description{
calculate iCMS on bulk data (microarray / RNA-seq)
}
