% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icms.ssc.R
\name{iCMS.DQ}
\alias{iCMS.DQ}
\title{calculate iCMS on bulk data (microarray / RNA-seq) using single-sample predictor}
\usage{
iCMS.DQ(
  ivect,
  min.cor = 0.1,
  min.dist = 0.05,
  q = 0.9,
  metric = "kendall",
  allgenes = FALSE,
  min.genes = 30,
  jobs = 4
)
}
\arguments{
\item{ivect}{input matrix to calculate. The row will be the gene symbol, and column will be sample name}

\item{min.cor}{the minimum correlation threshold to determine the confidence (default = 0.1)}

\item{min.dist}{the minimum distance to determine the iCMS call (default = 0.05)}

\item{q}{The quantile that is used to judge distance (default = 0.9). q=0.9-1 appears better for non batch and corrected data while q=0.6-0.7 appears better for batch-corrected}

\item{allgenes}{whether to use the extended genes or not, basically only epithelial genes are used (default = FALSE)}

\item{min.genes}{Minimum number of genes that can calculate iCMS call (default = 30 genes)}

\item{jobs}{If parallel execution fails you can try jobs=1 which will use serial apply (default = 4)}

\item{matric}{The method to calculate correlation matrix (default = kendall). Kendall correlation (non parametric) seems more robust but is slow. Pearson is faster, and often good enough.}
}
\value{
iCMS classification
}
\description{
calculate iCMS on bulk data (microarray / RNA-seq) using single-sample predictor
}
