% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_knn.R
\name{knn.vect}
\alias{knn.vect}
\title{calculate iCMS on bulk data (microarray / RNA-seq) using k-nearest neighbors algorithm for each sample}
\usage{
knn.vect(v, nn, metric, allgenes, min.genes, mcenter = FALSE, verbose = FALSE)
}
\arguments{
\item{v}{input matrix to calculate. The row will be the gene symbol, and column will be sample name}

\item{nn}{the minimum correlation threshold to determine the confidence (default = 0.1)}

\item{allgenes}{whether to use the extended genes or not, basically only epithelial genes are used (default = FALSE)}

\item{min.genes}{Minimum number of genes that can calculate iCMS call (default = 30 genes)}

\item{mcenter}{description}

\item{matric}{The method to calculate correlation matrix (default = kendall). Kendall correlation (non parametric) seems more robust but is slow. Pearson is faster, and often good enough.}
}
\value{
iCMS classification
}
\description{
calculate iCMS on bulk data (microarray / RNA-seq) using k-nearest neighbors algorithm for each sample
}
